fix: npe on delete node

+++ app/src/main/java/cc/cc1234/app/controller/NodeViewController.java
@@ -129,12 +129,17 @@ public class NodeViewController {
     }
 
     private void onNodeDelete() {
-        final String path = zkNodeTreeView.getSelectionModel().getSelectedItem().getValue().getPath();
-        Dialog.confirm("删除节点", "该操作将删除 " + path + " 该节点及其对应的子节点，操作不可恢复，请谨慎执行", () -> {
-            Try.of(() -> prettyZooFacade.deleteNode(ActiveServerContext.get(), path))
-                    .onFailure(exception -> VToast.error("delete failed:" + exception.getMessage()))
-                    .onSuccess(t -> VToast.info("delete success"));
-        });
+        final TreeItem<ZkNode> selectedItem = zkNodeTreeView.getSelectionModel().getSelectedItem();
+        if (selectedItem == null) {
+            VToast.error("select node first");
+        } else {
+            final String path = selectedItem.getValue().getPath();
+            Dialog.confirm("删除节点", "该操作将删除 " + path + " 该节点及其对应的子节点，操作不可恢复，请谨慎执行", () -> {
+                Try.of(() -> prettyZooFacade.deleteNode(ActiveServerContext.get(), path))
+                        .onFailure(exception -> VToast.error("delete failed:" + exception.getMessage()))
+                        .onSuccess(t -> VToast.info("delete success"));
+            });
+        }
     }
 
     private void initSearchTextField() {
